Add an untested pseudocode example of 1 button press = 1 action
The main loop executes VERY fast, usually faster than a human can press and release a button. Some things need to happen only once and not again until the human releases the button and presses it again. This is an example of a pattern to do that in a high-speed loop.